+++ libft/ft_lstnew.c
@@ -6,13 +6,13 @@
 /*   By: mel-aini <mel-aini@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 18:18:46 by mel-aini          #+#    #+#             */
-/*   Updated: 2023/02/09 17:25:00 by mel-aini         ###   ########.fr       */
+/*   Updated: 2023/02/25 17:51:50 by mel-aini         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-t_list	*ft_lstnew(int content)
+t_list	*ft_lstnew(long content)
 {
 	t_list	*head;
